requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
pandas>=1.2.0
scikit-learn>=0.24.0
Pillow>=8.0.0
tqdm>=4.60.0
matplotlib>=3.3.0
pyyaml>=5.4.0
timm>=0.9.0  # For ConvNeXtV2 and other advanced backbones
